+++ simpletransformers/config/model_args.py
@@ -100,19 +100,19 @@ class ClassificationArgs(ModelArgs):
     Model args for a ClassificationModel
     """
 
-    tie_value: int = 1
-    stride: float = 0.8
-    sliding_window: bool = False
-    regression: bool = False
-    lazy_text_column: int = 0
-    lazy_text_b_column: bool = None
-    lazy_text_a_column: bool = None
-    lazy_labels_column: int = 1
-    lazy_header_row: bool = True
-    lazy_delimiter: str = "\t"
     labels_list: list = field(default_factory=list)
     labels_map: dict = field(default_factory=dict)
+    lazy_delimiter: str = "\t"
+    lazy_labels_column: int = 1
     lazy_loading: bool = False
+    lazy_loading_start_line: int = 1
+    lazy_text_a_column: bool = None
+    lazy_text_b_column: bool = None
+    lazy_text_column: int = 0
+    regression: bool = False
+    sliding_window: bool = False
+    stride: float = 0.8
+    tie_value: int = 1
 
 
 @dataclass
@@ -202,6 +202,9 @@ class LanguageModelingArgs(ModelArgs):
     tie_generator_and_discriminator_embeddings: bool = True
     tokenizer_name: str = None
     vocab_size: int = None
+    clean_text: bool = True
+    handle_chinese_chars: bool = True
+    strip_accents: bool = True
     local_rank: int = -1
 
 
@@ -224,7 +227,7 @@ class Seq2SeqArgs(ModelArgs):
     repetition_penalty: float = 1.0
     top_k: float = None
     top_p: float = None
-    use_multiprocessed_decoding: bool = True
+    use_multiprocessed_decoding: bool = False
 
 
 @dataclass